Get meteorological stations data

Connect and retrieve data from AEMET, SMC and other Spanish meteorological stations services.

get_meteo_from(
  service = c("aemet", "meteocat", "meteoclimatic", "meteogalicia", "ria"),
  options
)

Arguments

service

Character with the service name (in lower case).

options

List with the needed service options. See services_options to have more info about the different services and their options.

Value

An sf (spatial) object with the stations meteorological data.

Details

Depending on the service and the temporal resolution selected, the variables present can change, but all services have at least temperature values.

API limits

Some APIs have limits in terms of the data that can be retrieved with one call. For example, AEMET only serves daily data for 31 days in one query. See vignette('api_limits', package = 'meteospain') for a detailed explanations of those limits and the ways to retrieve longer periods.

Cache

In order to avoid unnecessary API calls, results of this function are cached in memory with memoise. This means that subsequent calls from get_meteo_from with the same arguments will be faster as they will not call the meteorological service API. This cache has a maximum size of 1024 MB and persist 24 hours in the same R session after loading the package. To forgot the cache, memoise::forget(get_meteo_from) can be used.

Examples

# \donttest{
if (identical(Sys.getenv("NOT_CRAN"), "true")) {
  library(meteospain)
  library(keyring)

  # AEMET (we need a key)
  # key_set('aemet')
  options_for_aemet <- aemet_options(
    'daily',
    start_date = as.Date('2012-01-01'),
    end_date = as.Date('2012-02-01'),
    api_key = key_get('aemet')
  )
  get_meteo_from('aemet', options_for_aemet)
}
#> © AEMET. Autorizado el uso de la información y su reproducción citando a AEMET como autora de la misma.
#> https://www.aemet.es/es/nota_legal
#> Simple feature collection with 7529 features and 12 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -17.91528 ymin: 27.73583 xmax: 4.215556 ymax: 43.78611
#> Geodetic CRS:  WGS 84
#> # A tibble: 7,529 × 13
#>    timestamp           service station…¹ stati…² stati…³ altit…⁴ mean_…⁵ min_t…⁶
#>    <dttm>              <chr>   <chr>     <chr>   <chr>       [m]    [°C]    [°C]
#>  1 2012-01-01 00:00:00 aemet   0016A     "REUS … TARRAG…      71    11.4     4.3
#>  2 2012-01-01 00:00:00 aemet   0076      "BARCE… BARCEL…       4    12.9     8.7
#>  3 2012-01-01 00:00:00 aemet   0149X     "MANRE… BARCEL…     291     7.6     2.8
#>  4 2012-01-01 00:00:00 aemet   0200E     "BARCE… BARCEL…     408    13       9.4
#>  5 2012-01-01 00:00:00 aemet   0367      "GIRON… GIRONA      143    11.2     4.7
#>  6 2012-01-01 00:00:00 aemet   1002Y     "BAZTA… NAVARRA     183    12.4     8.7
#>  7 2012-01-01 00:00:00 aemet   1014      "HONDA… GIPUZK…       4    13.2     8.3
#>  8 2012-01-01 00:00:00 aemet   1014A     "SAN S… GIPUZK…       4    12.6     7.7
#>  9 2012-01-01 00:00:00 aemet   1024E     "DONOS… GIPUZK…     251    13.5    10.5
#> 10 2012-01-01 00:00:00 aemet   1037Y     "ZUMAR… GIPUZK…     420     8.8     4.2
#> # … with 7,519 more rows, 5 more variables: max_temperature [°C],
#> #   precipitation [L/m^2], mean_wind_speed [m/s], insolation [h],
#> #   geometry <POINT [°]>, and abbreviated variable names ¹​station_id,
#> #   ²​station_name, ³​station_province, ⁴​altitude, ⁵​mean_temperature,
#> #   ⁶​min_temperature
# }