Get meteorological stations info

Obtain info and metadata for the available stations in the different services

get_stations_info_from(
  service = c("aemet", "meteocat", "meteoclimatic", "meteogalicia"),
  options
)

Arguments

service
Character with the service name (in lower case).

options
List with the needed service options. See services_options to have more info about the different services and their options.

Value

An sf (spatial) object with the stations metadata.

Details

Depending on the service the metadata available can be different. Also, some services only offer info for active stations (i.e. AEMET), not historical stations, so some mismatch can occur between the stations returned by this function and the stations returned by get_meteo_from for historical dates.

Examples

# \donttest{
library(meteospain)
library(keyring)

# AEMET (we need a key)
# key_set('aemet')
api_options <- aemet_options(api_key = key_get('aemet'))
get_stations_info_from('aemet', api_options)

#> Simple feature collection with 291 features and 4 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -17.5455 ymin: 27.4409 xmax: 4.1256 ymax: 43.471
#> Geodetic CRS:  WGS 84
#> # A tibble: 291 × 5
#>    service station_id station_name           altitude         geometry
#>  * <chr>   <chr>      <chr>                       [m]      <POINT [°]>
#>  1 aemet   0252D      ARENYS DE MAR                74 (2.3224 41.3515)
#>  2 aemet   0076       BARCELONA AEROPUERTO          4 (2.0412 41.1734)
#>  3 aemet   0200E      BARCELONA, FABRA            408 (2.0727 41.2506)
#>  4 aemet   0201D      BARCELONA                     6   (2.12 41.2326)
#>  5 aemet   0149X      MANRESA                     291 (1.5025 41.4312)
#>  6 aemet   0229I      SABADELL AEROPUERTO         146 (2.0611 41.3125)
#>  7 aemet   0255B      SANTA SUSANNA                40 (2.4149 41.3903)
#>  8 aemet   0367       GIRONA AEROPUERTO           143 (2.4548 41.5442)
#>  9 aemet   0370B      GIRONA, ANTIC INSTITUT       95 (2.4931 41.5848)
#> 10 aemet   0372C      PORQUERES                   157 (2.4549 42.0616)
#> # … with 281 more rows

# }