Get meteorological stations data

Connect and retrieve data from AEMET, SMC and other spanish meteorological stations services.

get_meteo_from(
  service = c("aemet", "meteocat", "meteoclimatic", "meteogalicia"),
  options
)

Arguments

service
Character with the service name (in lower case).

options
List with the needed service options. See services_options to have more info about the different services and their options.

Value

An sf (spatial) object with the stations meteorological data.

Details

Depending on the service and the temporal resolution selected, the variables present can change, but all services have at least temperature values.

API limits

Some APIs have limits in terms of the data that can be retrieved with one call. For example, AEMET only serves daily data for 31 days in one query. See vignette('api_limits', package = 'meteospain') for a detailed explanations of those limits and the ways to retrieve longer periods.

Examples

# \donttest{
library(meteospain)
library(keyring)

# AEMET (we need a key)
# key_set('aemet')
options_for_aemet <- aemet_options(
  'daily',
  start_date = as.Date('2012-01-01'),
  end_date = as.Date('2012-02-01'),
  api_key = key_get('aemet')
)
get_meteo_from('aemet', options_for_aemet)

#> © AEMET. Autorizado el uso de la información y su reproducción citando a AEMET como autora de la misma.
#> http://www.aemet.es/es/nota_legal

#> Simple feature collection with 7529 features and 12 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -17.5455 ymin: 27.4409 xmax: 4.1256 ymax: 43.471
#> Geodetic CRS:  WGS 84
#> # A tibble: 7,529 × 13
#>    timestamp           service station_id station_name station_province altitude
#>    <dttm>              <chr>   <chr>      <chr>        <chr>                 [m]
#>  1 2012-01-01 00:00:00 aemet   0016A      "REUS AEROP… TARRAGONA              71
#>  2 2012-01-01 00:00:00 aemet   0076       "BARCELONA … BARCELONA               4
#>  3 2012-01-01 00:00:00 aemet   0149X      "MANRESA"    BARCELONA             291
#>  4 2012-01-01 00:00:00 aemet   0200E      "BARCELONA,… BARCELONA             408
#>  5 2012-01-01 00:00:00 aemet   0367       "GIRONA AER… GIRONA                143
#>  6 2012-01-01 00:00:00 aemet   1002Y      "BAZTAN, IR… NAVARRA               183
#>  7 2012-01-01 00:00:00 aemet   1014       "HONDARRIBI… GIPUZKOA                4
#>  8 2012-01-01 00:00:00 aemet   1014A      "SAN SEBAST… GIPUZKOA                4
#>  9 2012-01-01 00:00:00 aemet   1024E      "DONOSTIA/S… GIPUZKOA              251
#> 10 2012-01-01 00:00:00 aemet   1037Y      "ZUMARRAGA"  GIPUZKOA              420
#> # … with 7,519 more rows, and 7 more variables: mean_temperature [°C],
#> #   min_temperature [°C], max_temperature [°C], precipitation [L/m^2],
#> #   mean_wind_speed [m/s], insolation [h], geometry <POINT [°]>

# }